Move controllers and reconcile logic
This introduce reconciler pkg, this package provides more clarity
between controller and reconciler logic. It also allows each resource
to have an internal resources package to deal with all related resources
involved during that resource reconciliation.
