+++ pkg/reconciler/buildrun/buildrun_suite_test.go
@@ -1,0 +1,13 @@
+package buildrun_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestBuildrun(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Buildrun Suite")
+}
