+++ pkg/reconciler/buildrun/resources/runtime_image_test.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package buildrun
+package resources
 
 import (
 	"fmt"
