+++ pkg/reconciler/buildstrategy/controller.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package clusterbuildstrategy
+package buildstrategy
 
 import (
 	"context"
@@ -20,16 +20,16 @@ import (
 	"github.com/shipwright-io/build/pkg/ctxlog"
 )
 
-// Add creates a new ClusterBuildStrategy Controller and adds it to the Manager. The Manager will set fields on the Controller
+// Add creates a new BuildStrategy Controller and adds it to the Manager. The Manager will set fields on the Controller
 // and Start it when the Manager is Started.
 func Add(ctx context.Context, c *config.Config, mgr manager.Manager) error {
-	ctx = ctxlog.NewContext(ctx, "clusterbuildstrategy-controller")
+	ctx = ctxlog.NewContext(ctx, "buildstrategy-controller")
 	return add(ctx, mgr, NewReconciler(ctx, c, mgr))
 }
 
 // NewReconciler returns a new reconcile.Reconciler
 func NewReconciler(ctx context.Context, c *config.Config, mgr manager.Manager) reconcile.Reconciler {
-	return &ReconcileClusterBuildStrategy{
+	return &ReconcileBuildStrategy{
 		ctx:    ctx,
 		config: c,
 		client: mgr.GetClient(),
@@ -40,13 +40,13 @@ func NewReconciler(ctx context.Context, c *config.Config, mgr manager.Manager) r
 // add adds a new Controller to mgr with r as the reconcile.Reconciler
 func add(ctx context.Context, mgr manager.Manager, r reconcile.Reconciler) error {
 	// Create a new controller
-	c, err := controller.New("clusterbuildstrategy-controller", mgr, controller.Options{Reconciler: r})
+	c, err := controller.New("buildstrategy-controller", mgr, controller.Options{Reconciler: r})
 	if err != nil {
 		return err
 	}
 
-	// Watch for changes to primary resource ClusterBuildStrategy
-	err = c.Watch(&source.Kind{Type: &buildv1alpha1.ClusterBuildStrategy{}}, &handler.EnqueueRequestForObject{})
+	// Watch for changes to primary resource BuildStrategy
+	err = c.Watch(&source.Kind{Type: &buildv1alpha1.BuildStrategy{}}, &handler.EnqueueRequestForObject{})
 	if err != nil {
 		return err
 	}
@@ -54,27 +54,15 @@ func add(ctx context.Context, mgr manager.Manager, r reconcile.Reconciler) error
 	return nil
 }
 
-// blank assignment to verify that ReconcileClusterBuildStrategy implements reconcile.Reconciler
-var _ reconcile.Reconciler = &ReconcileClusterBuildStrategy{}
+// blank assignment to verify that ReconcileBuildStrategy implements reconcile.Reconciler
+var _ reconcile.Reconciler = &ReconcileBuildStrategy{}
 
-// ReconcileClusterBuildStrategy reconciles a ClusterBuildStrategy object
-type ReconcileClusterBuildStrategy struct {
+// ReconcileBuildStrategy reconciles a BuildStrategy object
+type ReconcileBuildStrategy struct {
 	// This client, initialized using mgr.Client() above, is a split client
 	// that reads objects from the cache and writes to the apiserver
 	ctx    context.Context
 	config *config.Config
 	client client.Client
 	scheme *runtime.Scheme
-}
-
-// Reconcile reads that state of the cluster for a ClusterBuildStrategy object and makes changes based on the state read
-// and what is in the BuildStrategy.Spec
-func (r *ReconcileClusterBuildStrategy) Reconcile(request reconcile.Request) (reconcile.Result, error) {
-
-	// Set the ctx to be Background, as the top-level context for incoming requests.
-	ctx, cancel := context.WithTimeout(r.ctx, r.config.CtxTimeOut)
-	defer cancel()
-
-	ctxlog.Info(ctx, "reconciling ClusterBuildStrategy", "name", request.Name)
-	return reconcile.Result{}, nil
 }
